backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
databricks-sql-connector==3.1.0
sqlglot==23.12.2
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
# Stage 3 – AI Orchestration (lazy-imported; mock mode works without it)
openai>=1.14.0
